+++ src/server/src/auth/findUser.ts
@@ -1,5 +1,7 @@
 import { User } from "../user.js";
+import { users } from "../../index.js";
+import { hashPassword } from "../auth/hash.js";
 
 export function findUserByPassword(password: string): User | void {
-    return 
+    return users.find((user: User) => user.passwordHash === hashPassword(password));
 }