+++ src/server/types/types.ts
@@ -4,5 +4,3 @@
 //     UUID: string;
 //     passwordHash: string;
 // }
-
-export {};